Conform to class-file and namespace conventions.
