+++ builds/msvc/vs2013/libbitcoin-server/libbitcoin-server.vcxproj
@@ -122,12 +122,24 @@
     <ClInclude Include="..\..\..\..\include\bitcoin\server\version.hpp" />
     <ClInclude Include="..\..\..\..\include\bitcoin\server\web\block_socket.hpp" />
     <ClInclude Include="..\..\..\..\include\bitcoin\server\web\heartbeat_socket.hpp" />
+    <ClInclude Include="..\..\..\..\include\bitcoin\server\web\http\bind_options.hpp" />
     <ClInclude Include="..\..\..\..\include\bitcoin\server\web\http\connection.hpp" />
+    <ClInclude Include="..\..\..\..\include\bitcoin\server\web\http\connection_state.hpp" />
+    <ClInclude Include="..\..\..\..\include\bitcoin\server\web\http\event.hpp" />
+    <ClInclude Include="..\..\..\..\include\bitcoin\server\web\http\file_transfer.hpp" />
     <ClInclude Include="..\..\..\..\include\bitcoin\server\web\http\http.hpp" />
+    <ClInclude Include="..\..\..\..\include\bitcoin\server\web\http\http_reply.hpp" />
+    <ClInclude Include="..\..\..\..\include\bitcoin\server\web\http\http_request.hpp" />
     <ClInclude Include="..\..\..\..\include\bitcoin\server\web\http\json_string.hpp" />
     <ClInclude Include="..\..\..\..\include\bitcoin\server\web\http\manager.hpp" />
+    <ClInclude Include="..\..\..\..\include\bitcoin\server\web\http\protocol_status.hpp" />
     <ClInclude Include="..\..\..\..\include\bitcoin\server\web\http\socket.hpp" />
+    <ClInclude Include="..\..\..\..\include\bitcoin\server\web\http\ssl.hpp" />
     <ClInclude Include="..\..\..\..\include\bitcoin\server\web\http\utilities.hpp" />
+    <ClInclude Include="..\..\..\..\include\bitcoin\server\web\http\websocket_frame.hpp" />
+    <ClInclude Include="..\..\..\..\include\bitcoin\server\web\http\websocket_message.hpp" />
+    <ClInclude Include="..\..\..\..\include\bitcoin\server\web\http\websocket_op.hpp" />
+    <ClInclude Include="..\..\..\..\include\bitcoin\server\web\http\websocket_transfer.hpp" />
     <ClInclude Include="..\..\..\..\include\bitcoin\server\web\query_socket.hpp" />
     <ClInclude Include="..\..\..\..\include\bitcoin\server\web\transaction_socket.hpp" />
     <ClInclude Include="..\..\..\..\include\bitcoin\server\workers\authenticator.hpp" />
