+++ include/bitcoin/server.hpp
@@ -37,12 +37,24 @@
 #include <bitcoin/server/web/heartbeat_socket.hpp>
 #include <bitcoin/server/web/query_socket.hpp>
 #include <bitcoin/server/web/transaction_socket.hpp>
+#include <bitcoin/server/web/http/bind_options.hpp>
 #include <bitcoin/server/web/http/connection.hpp>
+#include <bitcoin/server/web/http/connection_state.hpp>
+#include <bitcoin/server/web/http/event.hpp>
+#include <bitcoin/server/web/http/file_transfer.hpp>
 #include <bitcoin/server/web/http/http.hpp>
+#include <bitcoin/server/web/http/http_reply.hpp>
+#include <bitcoin/server/web/http/http_request.hpp>
 #include <bitcoin/server/web/http/json_string.hpp>
 #include <bitcoin/server/web/http/manager.hpp>
+#include <bitcoin/server/web/http/protocol_status.hpp>
 #include <bitcoin/server/web/http/socket.hpp>
+#include <bitcoin/server/web/http/ssl.hpp>
 #include <bitcoin/server/web/http/utilities.hpp>
+#include <bitcoin/server/web/http/websocket_frame.hpp>
+#include <bitcoin/server/web/http/websocket_message.hpp>
+#include <bitcoin/server/web/http/websocket_op.hpp>
+#include <bitcoin/server/web/http/websocket_transfer.hpp>
 #include <bitcoin/server/workers/authenticator.hpp>
 #include <bitcoin/server/workers/notification_worker.hpp>
 #include <bitcoin/server/workers/query_worker.hpp>
