+++ include/bitcoin/server/web/http/event.hpp
@@ -16,31 +16,27 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-#ifndef LIBBITCOIN_SERVER_WEB_HTTP_JSON_STRING_HPP
-#define LIBBITCOIN_SERVER_WEB_HTTP_JSON_STRING_HPP
+#ifndef LIBBITCOIN_SERVER_WEB_HTTP_EVENT_HPP
+#define LIBBITCOIN_SERVER_WEB_HTTP_EVENT_HPP
 
 #include <cstdint>
-#include <string>
-#include <bitcoin/server/define.hpp>
-#include <boost/property_tree/ptree.hpp>
-#include <bitcoin/bitcoin.hpp>
 
 namespace libbitcoin {
 namespace server {
 namespace http {
 
-// Object to JSON converters.
-//-----------------------------------------------------------------------------
-
-BCS_API std::string to_json(const boost::property_tree::ptree& tree);
-BCS_API std::string to_json(uint64_t height, uint32_t id);
-BCS_API std::string to_json(const code& code, uint32_t id);
-BCS_API std::string to_json(const chain::header& header, uint32_t id);
-BCS_API std::string to_json(const chain::block& block, uint32_t id);
-BCS_API std::string to_json(const chain::block& block, uint32_t height,
-    uint32_t id);
-BCS_API std::string to_json(const chain::transaction& transaction,
-    uint32_t id);
+enum class event : uint8_t
+{
+    read,
+    write,
+    listen,
+    accepted,
+    error,
+    closing,
+    websocket_frame,
+    websocket_control_frame,
+    json_rpc
+};
 
 } // namespace http
 } // namespace server
