+++ include/bitcoin/server/web/http/websocket_op.hpp
@@ -16,31 +16,24 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-#ifndef LIBBITCOIN_SERVER_WEB_HTTP_JSON_STRING_HPP
-#define LIBBITCOIN_SERVER_WEB_HTTP_JSON_STRING_HPP
+#ifndef LIBBITCOIN_SERVER_WEB_HTTP_WEBSOCKET_OP_HPP
+#define LIBBITCOIN_SERVER_WEB_HTTP_WEBSOCKET_OP_HPP
 
 #include <cstdint>
-#include <string>
-#include <bitcoin/server/define.hpp>
-#include <boost/property_tree/ptree.hpp>
-#include <bitcoin/bitcoin.hpp>
 
 namespace libbitcoin {
 namespace server {
 namespace http {
 
-// Object to JSON converters.
-//-----------------------------------------------------------------------------
-
-BCS_API std::string to_json(const boost::property_tree::ptree& tree);
-BCS_API std::string to_json(uint64_t height, uint32_t id);
-BCS_API std::string to_json(const code& code, uint32_t id);
-BCS_API std::string to_json(const chain::header& header, uint32_t id);
-BCS_API std::string to_json(const chain::block& block, uint32_t id);
-BCS_API std::string to_json(const chain::block& block, uint32_t height,
-    uint32_t id);
-BCS_API std::string to_json(const chain::transaction& transaction,
-    uint32_t id);
+enum class websocket_op : uint8_t
+{
+    continuation = 0,
+    text = 1,
+    binary = 2,
+    close = 8,
+    ping = 9,
+    pong = 10,
+};
 
 } // namespace http
 } // namespace server
